[ISSUE-2906] Send giphy extension
