+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/extensions/ChatClient.kt
@@ -230,3 +230,23 @@ internal fun ChatClient.needsMarkRead(cid: String): Boolean {
 
     return channelController.markRead()
 }
+
+/**
+ * Sends selected giphy message to the channel. Removes the original "ephemeral" message from local storage.
+ * Returns new "ephemeral" message with new giphy url.
+ * API call to remove the message is retried according to the retry policy specified on the chatDomain.
+ *
+ * @param message The message to send.
+ * @see io.getstream.chat.android.offline.utils.RetryPolicy
+ */
+internal fun ChatClient.sendGiphy(message: Message): Call<Message> {
+    val cid = message.cid
+    val domainImpl = domainImpl()
+    val channelController = domainImpl.channel(cid)
+
+    validateCid(cid)
+
+    return CoroutineCall(domainImpl.scope) {
+        channelController.sendGiphy(message)
+    }
+}
